Removing reading expressions from cmd-line
lc now *only* reads from stdin. This might change in the future, by
accepting a hypothetical -f option to pass it a file by filename
(although stdin redirection already works for that). This is to avoid
a problem with getopt(), namely, that it will *by design* always parse a
negative number as an option due to the leading - character.

README.md file updated accordingly.

Closes: #6